refactor generator stuff
