+++ src/KubeOps.Cli/Commands/Generator/Generate.cs
@@ -11,13 +11,7 @@ internal static class Generate
         {
             var cmd = new Command("generate", "Generates elements related to an operator.")
             {
-                CertificateGenerator.Command,
-                CrdGenerator.Command,
-                DockerGenerator.Command,
-                InstallerGenerator.Command,
                 OperatorGenerator.Command,
-                RbacGenerator.Command,
-                WebhookOperatorGenerator.Command,
             };
             cmd.AddAlias("gen");
             cmd.AddAlias("g");
