+++ src/KubeOps.Cli/Generators/DockerfileGenerator.cs
@@ -1,0 +1,33 @@
+using KubeOps.Cli.Output;
+
+namespace KubeOps.Cli.Generators;
+
+internal class DockerfileGenerator(bool hasWebhooks) : IConfigGenerator
+{
+    public void Generate(ResultOutput output)
+    {
+        output.Add(
+            "Dockerfile",
+            $"""
+             FROM mcr.microsoft.com/dotnet/sdk:latest as build
+             WORKDIR /operator
+
+             COPY ./ ./
+             RUN dotnet publish -c Release /p:AssemblyName=operator -o out
+
+             # The runner for the application
+             FROM mcr.microsoft.com/dotnet/{(hasWebhooks ? "aspnet" : "runtime")}:latest as final
+
+             RUN addgroup k8s-operator && useradd -G k8s-operator operator-user
+
+             WORKDIR /operator
+             COPY --from=build /operator/out/ ./
+             RUN chown operator-user:k8s-operator -R .
+
+             USER operator-user
+
+             ENTRYPOINT [ "dotnet", "operator.dll" ]
+             """,
+            OutputFormat.Plain);
+    }
+}
