+++ src/KubeOps.Cli/Generators/IConfigGenerator.cs
@@ -1,0 +1,10 @@
+using KubeOps.Cli.Output;
+
+using Spectre.Console;
+
+namespace KubeOps.Cli.Generators;
+
+internal interface IConfigGenerator
+{
+    void Generate(ResultOutput output);
+}
